+++ gbatchy/__init__.py
@@ -6,7 +6,7 @@ from .context import batch_context, BatchGreenlet, spawn, add_auto_wrapper, set_
 from .batch import batched, class_batched
 from .scheduler import Raise
 from .utils import (pmap, pmap_unordered, pfilter, pfilter_unordered, pget, immediate,
-                    immediate_exception, transform, spawn_proxy)
+                    immediate_exception, transform, spawn_proxy, iwait, wait, Pool)
 
 # Set up the default scheduler.
 from .scheduler import AllAtOnceScheduler as DefaultScheduler